Add tags finctionallity

+++ src/main/java/blog/controllers/PostController.java
@@ -2,13 +2,9 @@ package blog.controllers;
 
 import blog.forms.CommentForm;
 import blog.forms.PostForm;
-import blog.models.Comment;
-import blog.models.Post;
-import blog.models.User;
-import blog.services.CommentService;
-import blog.services.NotificationService;
-import blog.services.PostService;
-import blog.services.UserService;
+import blog.models.*;
+import blog.repositories.AuthUserDetailsService;
+import blog.services.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -21,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.validation.Valid;
 import javax.validation.constraints.AssertFalse;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Date;
@@ -36,13 +33,16 @@ public class PostController {
     private UserService userService;
     @Autowired
     private CommentService commentService;
-
+    @Autowired
+    private TagService tagService;
     @Autowired
     private NotificationService notificationService;
-
+    @Autowired
+    private AuthUserDetailsService authUserDetailsService;
 
     @RequestMapping("/posts")
     public String viewAllPost(Model model) {
+
         List<Post> posts = postService.findAll();
         if(posts == null) {
             notificationService.addErrorMessage("Cannot find any post");
@@ -80,27 +80,44 @@ public class PostController {
     }
 
     @RequestMapping(value = "/posts/delete/{id}", method = RequestMethod.POST)
-    public String deletePostById(@PathVariable("id") Long id) {
+    public String deletePostById(@PathVariable("id") Long id, @Valid PostForm postForm) {
         Post post = postService.findById(id);
         if(post == null) {
             notificationService.addErrorMessage("Cannot find  pos:" + id);
             return "redirect:/posts";
         }
-        // delete comment first
-        List<Comment> comments = commentService.findAllByPostId(id);
-        if (!comments.isEmpty()) {
-            for (Comment comment : comments) {
-                commentService.deleteById(comment.getId());
+        //check is user has a right
+
+        if(authUserDetailsService.isAuthorize() || postForm.getAuthorName().equals(post.getAuthor().getUsername()) ){
+
+            // delete comment first
+            List<Comment> comments = commentService.findAllByPostId(id);
+            if (!comments.isEmpty()) {
+                for (Comment comment : comments) {
+                    commentService.deleteById(comment.getId());
+                }
             }
+            //delete post_tags TODO
+            List<Post_tags> post_tags = tagService.findAllByPostId(id);
+            for (Post_tags tag : post_tags) {
+                if(tag.getPost().getId() == id){
+                   tagService.deletePost_tagsById(tag.getId());
+               }
+            }
+
+            postService.deleteById(id);
+            if(postService.findById(id) == null) {
+                notificationService.addInfoMessage("The post was deleted");
+                return "redirect:/posts";
+            } else {
+                notificationService.addErrorMessage("Cannot delete  pos:" + id);
+            }
+            return "posts/delete";
+
+        }else {
+            notificationService.addAlertMessage("There is not allowed to be here");
+            return "redirect:/";
         }
-        postService.deleteById(id);
-        if(postService.findById(id) == null) {
-            notificationService.addInfoMessage("The post was deleted");
-            return "redirect:/posts";
-        } else {
-            notificationService.addErrorMessage("Cannot delete  pos:" + id);
-        }
-        return "posts/delete";
     }
 
     @RequestMapping("/posts/edit/{id}")
@@ -110,21 +127,31 @@ public class PostController {
             notificationService.addErrorMessage("Cannot find  pos:" + id);
             return "redirect:/posts";
         }
-
+        List<Tag> tags = tagService.findAll();
+        model.addAttribute("tags", tags);
+        List<Post_tags> post_tagses  = tagService.findAllByPostId(id);
+        List<Tag> tags_old = new ArrayList<Tag>();
+        for (Post_tags post_tag: post_tagses ) {
+            tags_old.add(post_tag.getTag());
+        }
+        model.addAttribute("tags_old", tags_old);
         model.addAttribute("post", post);
         return "posts/edit";
     }
 
     @RequestMapping(value = "/posts/edit/{id}", method = RequestMethod.POST)
     public String editPost(@PathVariable("id") Long id, @Valid PostForm postForm, BindingResult bindingResult) {
+
         if (bindingResult.hasErrors()) {
             notificationService.addErrorMessage("Please fill all fields");
             return "posts/edit";
         }
         Post post = postService.findById(id);
+
+        //check is user has a right
+        if( authUserDetailsService.isAuthorize() || postForm.getAuthorName().equals(post.getAuthor().getUsername())){
         post.setTitle(postForm.getTitle());
         post.setBody(postForm.getBody());
-
 
         User author = userService.findByUsername(postForm.getAuthorName());
         post.setAuthor(author);
@@ -134,13 +161,36 @@ public class PostController {
 
         postService.edit(post);
 
-        notificationService.addInfoMessage("Post editted, Follow next step");
+        //delete old post_tags TODO
+            List<Post_tags> post_tags = tagService.findAllByPostId(id);
+            for (Post_tags tag : post_tags) {
+                if(tag.getPost().getId() == id){
+                    tagService.deletePost_tagsById(tag.getId());
+                }
+            }
+
+        //create new tags records
+        for (Tag tag : postForm.getTags()) {
+            Post_tags post_tagsNew = new Post_tags(tag, post);
+            tagService.create(post_tagsNew);
+        }
+
+
+
+            notificationService.addInfoMessage("Post editted, Follow next step");
         return "redirect:/posts";
+
+        }else {
+            notificationService.addAlertMessage("There is not allowed to be here");
+            return "redirect:/";
+        }
 
     }
 
     @RequestMapping("/posts/create")
-    public String showPostForm (PostForm postForm) {
+    public String showPostForm (PostForm postForm, Model model) {
+        List<Tag> tags = tagService.findAll();
+        model.addAttribute("tags", tags);
         return "posts/create";
     }
 
@@ -157,6 +207,12 @@ public class PostController {
 
         newPost.setDate(postForm.getDate());
         postService.create(newPost);
+        //create tags records
+        for (Tag tag : postForm.getTags()) {
+            Post_tags post_tags = new Post_tags(tag, newPost);
+            tagService.create(post_tags);
+        }
+
         if ( newPost.getId() > 0) {
             notificationService.addInfoMessage("Post created, Follow next step");
             return "redirect:/posts";
